spirit-tracks: finished refactoring new classes, need to update AP files
