+++ ItemClass.py
@@ -1,6 +1,6 @@
 
 from typing import TYPE_CHECKING
-from .subclasses import read_memory_value, split_bits
+from .subclasses import split_bits
 from ..data.Constants import DUNGEON_KEY_DATA
 
 if TYPE_CHECKING:
